Fix ridiculous load performance
Somehow I'd understood the node size of RBush to be the maximum points
and I was setting a massive number which gave terrible performance. Now
I've just set it to 16 which is what MapBox's Supercluster had it set at
before they transitioned to KDBush.
